Dynamically loading DApps & Modules from clients /modules subfolder through moduleLoader.js
