+++ clientSrc/launcher.html
@@ -7,10 +7,9 @@
     <body>
         <h3 style="text-align: center;">Seed Launcher</h3>
         <hr />
-        <p style="text-align: center;">
-            <input type="button" id="btnSeed" value="Seed" onclick="launch(&quot;Seed&quot;, &quot;seed.html&quot;)" />
-            <input type="button" id="btnRelay" value="Relay" onclick="launch(&quot;Relay&quot;, &quot;relay.html&quot;)" />
-        </p>
+        <div id="moduleButtons">
+        </div>
+        <input type="button" id="btnUnitTests" value="Run Unit Tests" onclick="runUnitTests()" />
     </body>
 </html>
