+++ clientSrc/modules/relay/index.html
@@ -2,7 +2,7 @@
 <html>
     <head>
         <link rel="stylesheet">
-        <script type="text/javascript" src="relay.js"></script>
+        <script type="text/javascript" src="index.js"></script>
     </head>
     <body>
         <center>
