+++ clientSrc/modules/seed/module.json
@@ -1,0 +1,6 @@
+{
+    "name" : "Seed",
+    "dappSource" : "index.html",
+    "moduleSource" : "seed.js",
+    "moduleChecksum" : "fd37"
+}
