+++ seedSrc/scenarioTest.js
@@ -16,8 +16,8 @@ const accountExporter = require("./account.js");
 const cryptographyHelper = cryptoHelperExporter.newCryptoHelper();
 const virtualMachineExporter = require("./virtualMachine/virtualMachine.js");
 const moduleExporter = require("./module.js");
-const seedExporter = require("./modules/seed.js");
-const relayExporter = require("./modules/relay.js");
+const seedExporter = require("../clientSrc/modules/seed/seed.js");
+const relayExporter = require("../clientSrc/modules/relay/relay.js");
 const moduleTester = require("./moduleTester.js");
 const transactionExporter = require("./transaction.js");
 const entanglementExporter = require("./entanglement.js");
@@ -37,6 +37,45 @@ let userChangeReply = function(payload) {
 }
 
 module.exports = {
+    seedScenarioSetupTest : function() {
+        console.log("### Seed Scenario Setup Test ###");
+        let tester = moduleTester.beginTest("Seed", "ABC", true);
+        tester.relay();
+        tester.relay();
+        tester.assertEqual("getBalanceOf", { owner : tester.getAccount("ABC") }, 1000, "Creator should start with 1000 SEED");
+        tester.assertEqual("getSymbol", {}, "SEED", "The symbol of Seed should be \"SEED\"");
+        tester.assertEqual("getDecimals", {}, 4, "Seed should have 4 decimal points");
+        tester.assertEqual("getTotalSupply", {}, 1000, "1000 SEED should be in circulation upon creation");
+        tester.assertEqual("getAllowance", { owner : tester.getAccount("ABC"), spender : tester.getAccount("DEF") }, undefined, "Get allowance is unset for user who has never used Seed before" );
+
+        tester.switchUser("ABC");
+        tester.createTransactionWithRelay("approve", { spender : tester.getAccount("DEF"), value : 250 });
+        tester.relay();
+
+        tester.switchUser("DEF");
+        tester.createTransactionWithRelay("transferFrom", { from : tester.getAccount("ABC"), to : tester.getAccount("DEF"), value : 100 });
+        tester.createTransactionWithRelay("transferFrom", { from : tester.getAccount("ABC"), to : tester.getAccount("GHI"), value : 100 });
+        tester.relay();
+        tester.relay();
+        tester.assertEqual("getBalanceOf", { owner : tester.getAccount("ABC") }, 800, "Owner should still have 800 SEED");
+        tester.assertEqual("getBalanceOf", { owner : tester.getAccount("DEF") }, 100, "DEF sent 100 SEED to himself");
+        tester.assertEqual("getBalanceOf", { owner : tester.getAccount("GHI") }, 100, "GHI received 100 SEED from DEF on ABC's behalf");
+
+        tester.switchUser("GHI");
+        tester.createTransactionWithRelay("transfer", { to : tester.getAccount("ABC"), value : 50 });
+        tester.relay();
+        tester.relay();
+        tester.assertEqual("getBalanceOf", { owner : tester.getAccount("ABC") }, 850, "ABC should have received 50 from GHI");
+
+        tester.switchUser("DEF");
+        tester.createTransactionWithRelay("burn", { value : 25 });
+        tester.relay();
+        tester.relay();
+        tester.assertEqual("getBalanceOf", { owner : tester.getAccount("DEF") }, 75, "DEF should have 75 after burning 25, removing it from circulation");
+        tester.assertEqual("getTotalSupply", {}, 975, "25 coins were burned, removed from circulation, since initial 1000 creation");
+
+        tester.endTest();
+    },
     seedAndSVMTransactionTest : function() {
         console.log("### Seed & SVM Transaction Test ###");
 
@@ -88,7 +127,7 @@ module.exports = {
 
         tester.endTest();
 
-        console.info("Entanglement", entanglementExporter.getEntanglement(), "Ledger", ledgerExporter.getLedger());
+        //console.info("Entanglement", entanglementExporter.getEntanglement(), "Ledger", ledgerExporter.getLedger());
         
     },
     seedAndSVMTransactionTest_NoDAG : function() {
@@ -137,7 +176,7 @@ module.exports = {
 
         tester.endTest();
 
-        console.info("Entanglement", entanglementExporter.getEntanglement(), "Ledger", ledgerExporter.getLedger().getModuleData("Seed"));
+        //console.info("Entanglement", entanglementExporter.getEntanglement(), "Ledger", ledgerExporter.getLedger().getModuleData("Seed"));
         
     },
     /**
