+++ seedSrc/virtualMachine/virtualMachine.js
@@ -55,9 +55,8 @@ class VirtualMachine {
      * Creates the module data in the ledger
      * 
      * @param {Module} newModule - An externally created Module object
-     * @param {string} creator - The public address of the creator of the module
      */
-    addModule(newModule, creator) {
+    addModule(newModule) {
         this.modules[newModule.module] = newModule;
         ledgerExporter.getLedger().addModuleData(newModule.module, newModule.initialData, newModule.initialUserData);
     }
@@ -116,6 +115,7 @@ class VirtualMachine {
         let moduleToInvoke = this.getModule(info);
         let moduleDataToInvoke = ledgerExporter.getLedger().getModuleData(info.module);
         //If the user simulating this does not exist, add them to our ledger
+        console.info("SIMULATE", moduleToInvoke, this);
         if (!moduleToInvoke.doesUserExist(info.user)) {
             this.addUser(info, info.user);
         }
@@ -198,7 +198,7 @@ class VirtualMachine {
      */
     createTransaction(account, mod, func, args, transactionsToValidate) {
         let tips = entanglement.getTipsToValidate(account.publicKey, transactionsToValidate);
-        console.info("createTransaction", account, mod, func, args, transactionsToValidate, tips);
+        console.info("createTransaction", account.publicKey, mod, func, args, transactionsToValidate);
         let localSimulation = this.simulate({ module : mod, function : func, args : args, user : account.publicKey, txHashes : tips });
         if (localSimulation.didChange()) {
             let work = this.doWork(account, tips);
